+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Fixed
 - Security - Fix potential vulnerability upgrading `acorn` dependency.
+- Technical - Unwrap default values in APIMap field builder.
 
 ## RELEASE 6.0.0 - 2020-03-17